factor (diagnosis): transaction isolation and code quality improvement
DiagnosisService의 여러 문제점을 해결하기 위해 리팩토링을 진행했습니다.

- **트랜잭션 자체 호출(self-invocation) 문제 해결**
  - `@Async` 메서드 내에서 호출되는 `@Transactional` 메서드들이 트랜잭션 프록시를 우회하여 적용되지 않는 문제를 해결했습니다.
  - `DiagnosisStateService`를 새로 생성하여 트랜잭션 관련 로직(상태 변경, 데이터 저장 및 로딩)을 분리했습니다.
  - `DiagnosisService`는 이제 `DiagnosisStateService`를 통해 트랜잭션 메서드를 호출하여 AOP 프록시가 정상적으로 동작하도록 보장합니다.

- **코드 가시성 및 경고 수정**
  - `public` 메서드에서 노출되던 `private` 레코드 `MemberWithProfile`을 별도의 `public` 파일로 분리하여 가시성 문제를 해결했습니다.
  - Lazy-loading을 위해 의도적으로 반환 값을 무시하던 코드에 `Objects.hashCode()`를 적용하여 "결과 무시" 경고를 제거하고 코드의 의도를 명확히 했습니다.

- **관련 변경 사항**
  - `DiagnosisService`에서 더 이상 직접 사용하지 않는 `MemberRepository` 의존성을 제거했습니다.
  - 사용되지 않는 `private` 메서드 및 `record` 정의를 삭제하여 코드 복잡성을 줄였습니다.
